+++ clang/docs/ReleaseNotes.rst
@@ -802,6 +802,11 @@ WebAssembly Support
 AVR Support
 ^^^^^^^^^^^
 
+SystemZ Support
+^^^^^^^^^^^^^^^
+
+- Add support for `#pragma export` and `_Export` keyword for z/OS
+
 DWARF Support in Clang
 ----------------------